requirements.txt
mcp
ollama
fastapi
uvicorn
pydantic
langchain
imap-tools
python-dateutil
pyda 
uvicorn 
psycopg2-binary 
sqlalchemy 
python-jose
passlib[bcrypt]
alembic
sqlmodel
pydantic-settings
python-json-logger
mangum